+++ star/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project aims to follow [Semantic Versioning](https://semver.org/).
 
 ---
 
-## [Unreleased]
+## [0.1.27] 2026-07-14
 
 ### 🐛 Fixed